backend/requirements.txt
Flask
Flask-Cors
mysql-connector-python
SQLAlchemy
tensorflow
pymysql
scikit-learn

